perf: lazy load file preview modules
